Added Normalizer object
The Normalizer object is needed to do one job: To normalize things.

Initially, there had existed HttpRequest->normalizePath(), but there
lack a path normalizer for the route paths. This meant that URI paths
from the client were normalized to adhere to the designed rules, but the
paths configured by the app developer were not adhering to the same
rules. This meant that there existed an incompatibility between the URI
path and the route path. So, in order to apply the rules across both
paths, the Normalizer object was created to enforce uniformity among
both paths.

The Normalizer was also created to be a general normalizer, and not just
to normalize URI-like paths.